rename: Icon/MaterialSymbols.tsx -> Icons/MaterialSymbols.tsx
